Logins, image uploads for authenticated users. Register new users.
Because GitHub does not want Sql Database files to be added to source
control, this will first fine out if the database is there. If the file
is present, it verifies the non authentication table has been entered.
If it has not, it prompts the user to make some file renames. If the
table is missing, it prompts the creation of a new user, and then
creates the image file table.